requirements.txt
# WSGI server
gunicorn

# Flask and assorted libraries
flask==1.*
flask-satella-metrics
flask-requests-logging>=0.5
flask-cors
flask-json
flasgger
pyjwt
markupsafe==2.0.*
cassandra-driver
pyyaml
satella